backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.0
fastapi-utils==0.2.1
h11==0.14.0
idna==3.6
joblib==1.3.2
numpy==1.26.4
pandas==2.2.1
pydantic==1.10.14
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.36.3
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
